refactor: move entity classes to the com.college.entity package
