+++ src/integration-test/java/com/college/controller/EnrollmentControllerIntegrationTest.java
@@ -1,6 +1,9 @@
 package com.college.controller;
 
-import com.college.*;
+import com.college.MainApp;
+import com.college.entity.Course;
+import com.college.entity.Department;
+import com.college.entity.Student;
 import com.college.service.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
