+++ src/main/java/com/college/MainApp.java
@@ -5,7 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 
 @SpringBootApplication
-@EntityScan(basePackages = {"com.college"})
+@EntityScan(basePackages = {"com.college.entity"})
 public class MainApp {
     public static void main(String[] args) {
         SpringApplication.run(MainApp.class, args);
