+++ src/main/java/com/college/entity/ClassSchedule.java
@@ -1,4 +1,4 @@
-package com.college;
+package com.college.entity;
 
 import javax.persistence.*;
 import java.time.LocalTime;
