+++ src/main/java/com/college/entity/Department.java
@@ -1,4 +1,4 @@
-package com.college;
+package com.college.entity;
 
 import javax.persistence.*;
