+++ src/main/java/com/college/entity/ScheduleInfoDTO.java
@@ -1,4 +1,4 @@
-package com.college;
+package com.college.entity;
 
 import java.time.LocalTime;
