+++ src/main/java/com/college/repository/ClassScheduleRepository.java
@@ -1,8 +1,9 @@
 package com.college.repository;
 
-import com.college.ClassSchedule;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.college.entity.ClassSchedule;
 
 @Repository
 public interface ClassScheduleRepository extends JpaRepository<ClassSchedule, Long> {
