+++ src/main/java/com/college/repository/CourseRepository.java
@@ -1,8 +1,9 @@
 package com.college.repository;
 
-import com.college.Course;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.college.entity.Course;
 
 @Repository
 public interface CourseRepository extends JpaRepository<Course, Long> {
