+++ src/main/java/com/college/repository/DepartmentRepository.java
@@ -1,6 +1,6 @@
 package com.college.repository;
 
-import com.college.Department;
+import com.college.entity.Department;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
