+++ src/main/java/com/college/repository/EnrollmentRepository.java
@@ -1,8 +1,9 @@
 package com.college.repository;
 
-import com.college.Enrollment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.college.entity.Enrollment;
 
 @Repository
 public interface EnrollmentRepository extends JpaRepository<Enrollment, Long> {
