+++ src/main/java/com/college/repository/RoomRepository.java
@@ -1,8 +1,9 @@
 package com.college.repository;
 
-import com.college.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.college.entity.Room;
 
 @Repository
 public interface RoomRepository extends JpaRepository<Room, Long> {
