+++ src/main/java/com/college/repository/StudentRepository.java
@@ -1,8 +1,9 @@
 package com.college.repository;
 
-import com.college.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.college.entity.Student;
 
 @Repository
 public interface StudentRepository extends JpaRepository<Student, Long> {
