+++ src/main/java/com/college/repository/TeacherRepository.java
@@ -1,8 +1,9 @@
 package com.college.repository;
 
-import com.college.Teacher;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.college.entity.Teacher;
 
 @Repository
 public interface TeacherRepository extends JpaRepository<Teacher, Long> {
