+++ src/main/java/com/college/service/DepartmentService.java
@@ -1,6 +1,6 @@
 package com.college.service;
 
-import com.college.Department;
+import com.college.entity.Department;
 import com.college.repository.DepartmentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
